+++ myqueue/pbs.py
@@ -9,9 +9,12 @@ from .scheduler import Scheduler
 
 
 class PBS(Scheduler):
+    output_file_pattern = r'[a-zA-Z]+.*\.[eo]+[1-9]+[0-9]*'
+
     def submit(self,
                task: Task,
-               dry_run: bool = False) -> None:
+               dry_run: bool = False,
+               verbose: bool = False) -> None:
         nodelist = config['nodes']
         nodes, nodename, nodedct = task.resources.select(nodelist)